user-vehicle-booking-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
Django==4.1.13
djangorestframework==3.14.0
dnspython==2.6.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
nodeenv==1.8.0
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.2
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.0
sqlparse==0.2.4
StrEnum==0.4.15
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
